std::list implementation changed, as it is in CSR

+++ include/ALGraph.h
@@ -38,7 +38,7 @@ public:
 
     void sortByEdgesByNodeId();
 
-    inline std::list<int>::iterator begin(int cur_vertex) {
+    inline std::list<uint64_t>::iterator begin(int cur_vertex) {
         return edges[cur_vertex].begin();
     }
